feat: do deploy

See merge request s12-webmobile2-sub1/S12P11A707!78

+++ backend/ttabong/src/main/java/com/ttabong/dto/sns/ReviewCommentDto.java
@@ -1,9 +1,10 @@
 package com.ttabong.dto.sns;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
 import java.time.LocalDateTime;
 
 @Getter
@@ -11,9 +12,9 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class ReviewCommentDto {
-    private Long commentId;
-    private Long writerId;
-    private Long reviewId;
+    private Integer commentId;
+    private Integer writerId;
+    private Integer reviewId;
     private String content;
     private Boolean isDeleted;
     private LocalDateTime updatedAt;
