+++ backend/ttabong/src/main/java/com/ttabong/dto/sns/ReviewDto.java
@@ -1,9 +1,10 @@
 package com.ttabong.dto.sns;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
 import java.time.LocalDateTime;
 
 @Getter
@@ -11,12 +12,12 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class ReviewDto {
-    private Long reviewId;         // 후기 ID
-    private Long parentReviewId;   // 자기참조 (부모 리뷰 ID)
-    private Long groupId;          // 그룹 ID
-    private Long recruitId;        // 공고 ID
-    private Long orgId;            // 작성하는 기관 ID
-    private Long writerId;         // 작성자 ID
+    private Integer reviewId;         // 후기 ID
+    private Integer parentReviewId;   // 자기참조 (부모 리뷰 ID)
+    private Integer groupId;          // 그룹 ID
+    private Integer recruitId;        // 공고 ID
+    private Integer orgId;            // 작성하는 기관 ID
+    private Integer writerId;         // 작성자 ID
     private String title;          // 후기 제목
     private String content;        // 후기 내용
     private Boolean isDeleted;     // 삭제 여부
