+++ backend/ttabong/src/main/java/com/ttabong/dto/sns/ReviewImageDto.java
@@ -1,9 +1,10 @@
 package com.ttabong.dto.sns;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
 import java.time.LocalDateTime;
 
 @Getter
@@ -11,10 +12,10 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class ReviewImageDto {
-    private Long imageId;
-    private Long reviewId;
+    private Integer imageId;
+    private Integer reviewId;
     private String imageUrl;
     private Boolean isDeleted;
     private LocalDateTime createdAt;
-    private Long nextImageId; // 다음 이미지 ID
+    private Integer nextImageId; // 다음 이미지 ID
 }
