+++ backend/ttabong/src/main/java/com/ttabong/dto/sns/response/CommentDeleteResponseDto.java
@@ -1,18 +1,15 @@
 package com.ttabong.dto.sns.response;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
-/*
- * 11. 댓글 _ 삭제
- * */
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class CommentDeleteResponseDto {
     private String message;
-    private Long commentId;
+    private Integer commentId;
 }
