+++ backend/ttabong/src/main/java/com/ttabong/dto/sns/response/ReviewDeleteResponseDto.java
@@ -1,20 +1,17 @@
 package com.ttabong.dto.sns.response;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
-/*
- * 6. 후기 _ 삭제
- * */
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class ReviewDeleteResponseDto {
     private String message;
-    private Long reviewId;
+    private Integer reviewId;
     private String title;
     private String content;
 }
