+++ backend/ttabong/src/main/java/com/ttabong/dto/sns/response/ReviewUpdateResponseDto.java
@@ -1,19 +1,16 @@
 package com.ttabong.dto.sns.response;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
-/*
- * 8. 후기 _ 수정
- * */
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class ReviewUpdateResponseDto {
     private String message;
-    private Long reviewId;
-    private Long writerId;
+    private Integer reviewId;
+    private Integer writerId;
 }
