+++ backend/ttabong/src/main/java/com/ttabong/dto/sns/response/ReviewVisibilitySettingResponseDto.java
@@ -1,21 +1,19 @@
 package com.ttabong.dto.sns.response;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
 import java.time.LocalDateTime;
 
-/*
- * 5. 후기 _ 공개 여부 설정
- * */
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class ReviewVisibilitySettingResponseDto {
     private String message;
-    private Long reviewId;
+    private Integer reviewId;
     private Boolean isPublic;
     private LocalDateTime updatedAt;
 }
